refactor: renamed fixedDimensions in tests
