+++ src/modules/layout/__tests__/layoutReplaceDimension.spec.js
@@ -1,6 +1,6 @@
 import { layoutReplaceDimension } from '../layoutReplaceDimension'
 import { TEST_LAYOUT, TEST_DIMENSION_4 } from '../testResources'
-import { DIMENSION_ID_ORGUNIT } from '../../fixedDimensions'
+import { DIMENSION_ID_ORGUNIT } from '../../predefinedDimensions'
 
 const newOuDimensions = [
     {
